backend/requirements.txt
# Django core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Authentication & Security
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5

# File handling & validation
Pillow==10.1.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.25

# HTTP requests
requests==2.31.0
httpx==0.25.2

# WhatsApp integration
# Removido pacote indisponível; integrações via HTTP serão feitas com requests/httpx

# AI/ML for triage
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# Cache
redis==5.0.1
django-redis==5.4.0

# Background tasks
celery==5.3.4
kombu==5.3.4

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Development & debugging
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Utilities
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1

# Health checks
django-health-check==3.17.0

# Logging
structlog==23.2.0
